move complex collection test to test_collection
