+++ tests/plotting/test_collection.py
@@ -1,4 +1,6 @@
 """This module contains any tests which cause memory leaks."""
+import weakref
+
 import numpy as np
 
 
@@ -13,3 +15,20 @@ def test_pyvistandarray_assign(sphere):
 def test_pyvistandarray_strides(sphere):
     sphere['test_scalars'] = sphere.points[:, 2]
     assert np.allclose(sphere['test_scalars'], sphere.points[:, 2])
+
+
+def test_complex_collection(plane):
+    name = 'my_data'
+    data = np.random.random((plane.n_points, 2)).view(np.complex128).ravel()
+    plane.point_data[name] = data
+
+    # ensure shallow copy
+    assert np.shares_memory(plane.point_data[name], data)
+
+    # ensure data remains but original numpy object does not
+    ref = weakref.ref(data)
+    data_copy = data.copy()
+    del data
+    assert np.allclose(plane.point_data[name], data_copy)
+
+    assert ref() is None
